requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
Deprecated==1.2.14
distro==1.9.0
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
netaddr==1.3.0
npm==0.1.1
optional-django==0.1.0
py3dns==4.0.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
requests==2.32.3
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
theblockchainapi==0.0.604
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
validate-email==1.3
validate-email-address==1
wrapt==1.16.0
yookassa==3.1.0
zope.interface==6.4.post2
solana~=0.34.3